+++ src/ide/init-intellij-settings.ts
@@ -3,25 +3,41 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as xml2js from 'xml2js';
 
-export interface SingleModifier {
+export interface SingleCommonConfig {
   name: string
   defaultXMLPath: string
   targetXMLPath: string
-  validator(json:[]|object): boolean
-  modifier(json:[]|object): string
+  replaceIfInvalid?: boolean
 }
 
-export async function initIntellijSettings(modifiers:SingleModifier[]) {
+export interface SingleModifier extends SingleCommonConfig {
+  validator(parsedJson:any): boolean
+  modifier(parsedJson:any): string
+  replaceIfExists: false
+}
+export interface SingleReplacer extends SingleCommonConfig{
+  replaceIfExists: true
+}
+
+export type SingleModifierOrReplacer = SingleModifier|SingleReplacer
+export type ListOfModifiersOrReplacers = SingleModifierOrReplacer[]
+
+export async function initIntellijSettings(modifiersOrReplacers:ListOfModifiersOrReplacers) {
   try {
     shell.clear();
     shell.newLine();
     shell.write(chalk.bold('Initialize Intellij Settings'));
     shell.newLine();
 
-    for (const modifierConfig of modifiers) {
+    for (const config of modifiersOrReplacers) {
       shell.newLine()
-      shell.write('  ' + modifierConfig.name)
-      await modifyXMLSettings(modifierConfig)
+      shell.write('  ' + config.name)
+
+      config.replaceIfExists
+        ? await replaceXMLSettings(config as SingleReplacer)
+        : await modifyXMLSettings(config as SingleModifier)
+
+      shell.write(chalk.green(config.targetXMLPath));
     }
 
     shell.newLine();
@@ -39,40 +55,72 @@ export async function initIntellijSettings(modifiers:SingleModifier[]) {
   }
 }
 
+async function replaceXMLSettings (config:SingleModifierOrReplacer) {
+  const { targetXMLPath } = config
+  const { defaultXMLString, targetXMLExists } = prepareForChangingSettings(config)
+
+  if (targetXMLExists)
+    fs.rmSync(targetXMLPath)
+
+  fs.writeFileSync(targetXMLPath, defaultXMLString);
+
+  targetXMLExists
+    ? shell.write(chalk.green.bold('    replaced '))
+    : shell.write(chalk.green.bold('    created '));
+}
+
 async function modifyXMLSettings (config:SingleModifier) {
-  const {
-    defaultXMLPath,
-    targetXMLPath,
-    validator,
-    modifier,
-  } = config
+  const { targetXMLPath, validator, modifier } = config
+  const { defaultXMLString, targetXMLNotExists } = prepareForChangingSettings(config)
 
-  if (fs.existsSync(defaultXMLPath) === false) throw new Error(`File "${defaultXMLPath}" must exists.`);
+  try {
+    if (targetXMLNotExists) {
+      fs.writeFileSync(targetXMLPath, defaultXMLString);
+      shell.write(chalk.green.bold('    created '));
+    }
+    else {
+      const targetXML = fs.readFileSync(targetXMLPath, 'utf-8');
+      const targetJSON = await xml2js.parseStringPromise(targetXML);
 
-  const defaultXMLString = fs.readFileSync(defaultXMLPath, 'utf-8');
+      if (targetJSON === null || !validator(targetJSON))
+        throw new Error(`File "${targetXMLPath}" has unexpected structure or it's corrupted.`);
+
+      const modifiedTargetJSON = modifier(targetJSON);
+
+      const builder = new xml2js.Builder();
+      const newTargetXML = builder.buildObject(modifiedTargetJSON);
+
+      fs.writeFileSync(targetXMLPath, newTargetXML);
+      shell.write(chalk.green.bold('    modified '));
+    }
+  }
+  catch (err) {
+    if (config.replaceIfInvalid === true)
+      await replaceXMLSettings(config)
+    else
+      throw err
+  }
+}
+
+function prepareForChangingSettings (config:SingleModifierOrReplacer) {
+  const { defaultXMLPath, targetXMLPath } = config
+
+  if (!fs.existsSync(defaultXMLPath)) throw new Error(`File "${defaultXMLPath}" must exists.`);
+
   const targetXMLExists = fs.existsSync(targetXMLPath);
-  const dirnameTarget = path.dirname(targetXMLPath);
+  const targetXMLNotExists = !targetXMLExists;
+  const defaultXMLString = fs.readFileSync(defaultXMLPath, 'utf-8');
+  const dirnameTargetPath = path.dirname(targetXMLPath);
+  const dirnameTargetExists = fs.existsSync(dirnameTargetPath);
+  const dirnameTargetNotExists = !dirnameTargetExists;
 
-  if (targetXMLExists === false) {
-    fs.existsSync(dirnameTarget) === false && fs.mkdirSync(dirnameTarget, { recursive: true });
-    fs.writeFileSync(targetXMLPath, defaultXMLString);
+  dirnameTargetNotExists && fs.mkdirSync(dirnameTargetPath, { recursive: true });
+
+  return {
+    defaultXMLString,
+    targetXMLExists,
+    targetXMLNotExists,
   }
-  else {
-    const targetXML = fs.readFileSync(targetXMLPath, 'utf-8');
-
-    const targetJSON = await xml2js.parseStringPromise(targetXML);
-    if (targetJSON === null || validator(targetJSON)) throw new Error(`File "${targetXMLPath}" has invalid structure or is corrupted.`);
-
-    const modifiedTargetJSON = modifier(targetJSON);
-
-    const builder = new xml2js.Builder();
-    const newTargetXML = builder.buildObject(modifiedTargetJSON);
-
-    fs.writeFileSync(targetXMLPath, newTargetXML);
-  }
-
-  shell.write(chalk.green.bold('    done '));
-  shell.write(chalk.green(targetXMLPath));
 }
 
 const shell = {