+++ CHANGELOG.md
@@ -6,12 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.3.2] – 2022-09-20
+## [0.4.0] – 2022-09-20
+### Updated
+- IDE: Docs for init-intellij-settings
+- IDE: Plugin API refactor
+### Added
+- IDE: Possible to replace target file if exists 
+- IDE: Possible to replace target file if it has unexpected structure, or it's corrupted
 ### Fixed
-- IDE: Add default-validators and modifiers
-- IDE: Add xml-utils
-- IDE: Show errors stack
+- IDE: Validator condition
+
+## [0.3.2] – 2022-09-20
+### Added
+- IDE: default-validators and modifiers
+- IDE: xml-utils
 - IDE: Add default xml configuration files
+### Fixed
+- IDE: Show errors stack
 
 ## [0.3.1] – 2022-09-19
 ### Fixed